daphnia_task: Fix failed consecutive notifies bug
When handle_info in a task is called, it may yield a new notify.
In this case, we can't progress the task state until we've called
handle_task_notify/2. If the task is set to proceed, we should call
maybe_progress/1 on the new state, otherwise we can pass the reply up.